requirements.txt
aenum==2.2.1
certifi==2019.6.16
chainer==6.3.0
chardet==3.0.4
client==0.0.1
filelock==3.0.12
gremlinpython==3.4.3
idna==2.8
isodate==0.6.0
Janome==0.3.9
joblib==0.13.2
jpholiday==0.0.6
mecab-python3==0.996.2
numpy==1.17.1
protobuf==3.7.1
pytz==2019.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
Serializer==0.2.1
simplejson==3.16.0
six==1.12.0
sklearn==0.0
tornado==4.5.3
typing==3.6.6
typing-extensions==3.6.6
urllib3==1.25.3
